refactor Modal and form component relates #9
